requirements.txt
absl-py==2.1.0
accelerate==1.1.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
astunparse==1.6.3
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.0
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.12.1
huggingface-hub==0.25.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.6.17
keras==3.8.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
munch==4.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.2
peft==0.13.2
pillow==10.4.0
pretrainedmodels==0.7.4
protobuf==5.28.1
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pyparsing==3.1.4
PySnooper==1.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch_toolbelt==0.8.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scalesim==2.0.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
shapely==2.0.6
simsimd==6.2.1
six==1.16.0
stringzilla==3.11.0
sympy==1.13.2
tbb==2021.11.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.9.7
tokenizers==0.20.3
torch==2.4.1
torchao==0.8.0
torchinfo==1.8.0
torchmetrics==1.7.1
TorchSnooper==0.8
torchvision==0.19.1
tqdm==4.66.5
transformers==4.46.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.17.2
yacs==0.1.8
zipp==3.20.2
